+++ arcade/gui/ui_manager.py
@@ -12,11 +12,21 @@ from collections import defaultdict
 from typing import Iterable, TypeVar, Union
 
 from pyglet.event import EVENT_HANDLED, EVENT_UNHANDLED, EventDispatcher
+from pyglet.input import Controller
+from pyglet.math import Vec2
 from typing_extensions import TypeGuard
 
 import arcade
+from arcade.experimental.controller_window import ControllerWindow
 from arcade.gui import UIEvent
 from arcade.gui.events import (
+    UIControllerButtonPressEvent,
+    UIControllerButtonReleaseEvent,
+    UIControllerConnectEvent,
+    UIControllerDisconnectEvent,
+    UIControllerDpadEvent,
+    UIControllerStickEvent,
+    UIControllerTriggerEvent,
     UIKeyPressEvent,
     UIKeyReleaseEvent,
     UIMouseDragEvent,
@@ -278,6 +288,20 @@ class UIManager(EventDispatcher):
         """
         if not self._enabled:
             self._enabled = True
+
+            if isinstance(self.window, ControllerWindow):
+                controller_handlers = {
+                    self.on_connect,
+                    self.on_disconnect,
+                    self.on_stick_motion,
+                    self.on_trigger_motion,
+                    self.on_button_press,
+                    self.on_button_release,
+                    self.on_dpad_motion,
+                }
+            else:
+                controller_handlers = set()
+
             self.window.push_handlers(
                 self.on_resize,
                 self.on_update,
@@ -291,6 +315,7 @@ class UIManager(EventDispatcher):
                 self.on_text,
                 self.on_text_motion,
                 self.on_text_motion_select,
+                *controller_handlers,
             )
 
     def disable(self) -> None:
@@ -301,6 +326,20 @@ class UIManager(EventDispatcher):
         """
         if self._enabled:
             self._enabled = False
+
+            if isinstance(self.window, ControllerWindow):
+                controller_handlers = {
+                    self.on_connect,
+                    self.on_disconnect,
+                    self.on_stick_motion,
+                    self.on_trigger_motion,
+                    self.on_button_press,
+                    self.on_button_release,
+                    self.on_dpad_motion,
+                }
+            else:
+                controller_handlers = set()
+
             self.window.remove_handlers(
                 self.on_resize,
                 self.on_update,
@@ -314,6 +353,7 @@ class UIManager(EventDispatcher):
                 self.on_text,
                 self.on_text_motion,
                 self.on_text_motion_select,
+                *controller_handlers,
             )
 
     def on_update(self, time_delta):
@@ -449,6 +489,29 @@ class UIManager(EventDispatcher):
             surface.resize(size=(width, height), pixel_ratio=scale)
 
         self.trigger_render()
+
+    def on_connect(self, controller: Controller):
+        """Called when a controller is connected."""
+        self.dispatch_ui_event(UIControllerConnectEvent(controller))
+
+    def on_disconnect(self, controller: Controller):
+        """Called when a controller is disconnected."""
+        self.dispatch_ui_event(UIControllerDisconnectEvent(controller))
+
+    def on_stick_motion(self, controller: Controller, name: str, value: Vec2):
+        return self.dispatch_ui_event(UIControllerStickEvent(controller, name, value))
+
+    def on_trigger_motion(self, controller: Controller, name: str, value: float):
+        return self.dispatch_ui_event(UIControllerTriggerEvent(controller, name, value))
+
+    def on_button_press(self, controller: Controller, button: str):
+        return self.dispatch_ui_event(UIControllerButtonPressEvent(controller, button))
+
+    def on_button_release(self, controller: Controller, button: str):
+        return self.dispatch_ui_event(UIControllerButtonReleaseEvent(controller, button))
+
+    def on_dpad_motion(self, controller: Controller, value: Vec2):
+        return self.dispatch_ui_event(UIControllerDpadEvent(controller, value))
 
     @property
     def rect(self) -> Rect: