+++ arcade/gui/widgets/__init__.py
@@ -1,5 +1,8 @@
+from __future__ import annotations
+
 from abc import ABC
-from typing import Dict, Iterable, List, NamedTuple, Optional, TYPE_CHECKING, Tuple, TypeVar, Union
+from enum import IntEnum
+from typing import TYPE_CHECKING, Dict, Iterable, List, NamedTuple, Optional, Tuple, TypeVar, Union
 
 from pyglet.event import EVENT_HANDLED, EVENT_UNHANDLED, EventDispatcher
 from pyglet.math import Vec2
@@ -25,9 +28,20 @@ from arcade.utils import copy_dunders_unimplemented
 if TYPE_CHECKING:
     from arcade.gui.ui_manager import UIManager
 
-__all__ = ["Surface", "UIDummy"]
-
 W = TypeVar("W", bound="UIWidget")
+
+
+class FocusMode(IntEnum):
+    """Defines the focus mode of a widget.
+
+    0: Not focusable
+    1: Focusable
+
+    We might support different focus modes in the future, but for now on/off is enough.
+    """
+
+    NONE = 0
+    ALL = 2
 
 
 class _ChildEntry(NamedTuple):
@@ -57,6 +71,8 @@ class UIWidget(EventDispatcher, ABC):
 
     rect = Property(LBWH(0, 0, 1, 1))
     visible = Property(True)
+    focused = Property(False)
+    focus_mode: FocusMode = FocusMode.NONE
 
     size_hint = Property[Optional[Tuple[Optional[float], Optional[float]]]](None)
     size_hint_min = Property[Optional[Tuple[Optional[float], Optional[float]]]](None)
@@ -107,6 +123,7 @@ class UIWidget(EventDispatcher, ABC):
             self.add(child)
 
         bind(self, "rect", self.trigger_full_render)
+        bind(self, "focused", self.trigger_full_render)
         bind(
             self, "visible", self.trigger_full_render
         )  # TODO maybe trigger_parent_render would be enough
@@ -242,6 +259,8 @@ class UIWidget(EventDispatcher, ABC):
             rendered = True
             self.do_render_base(surface)
             self.do_render(surface)
+            if self.focused:
+                self.do_render_focus(surface)
             self._requires_render = False
 
         # only render children if self is visible
@@ -292,6 +311,15 @@ class UIWidget(EventDispatcher, ABC):
         """
         pass
 
+    def do_render_focus(self, surface: Surface):
+        """Render the widgets focus representation overlay`"""
+        self.prepare_render(surface)
+        arcade.draw_rect_outline(
+            rect=LBWH(0, 0, self.content_width, self.content_height),
+            color=arcade.color.WHITE,
+            border_width=4,
+        )
+
     def dispatch_ui_event(self, event: UIEvent):
         """Dispatch a :class:`UIEvent` using pyglet event dispatch mechanism"""
         return self.dispatch_event("on_event", event)
@@ -313,6 +341,19 @@ class UIWidget(EventDispatcher, ABC):
             anchor: anchor point
         """
         self.rect = self.rect.scale(new_scale=factor, anchor=anchor)
+
+    def get_ui_manager(self) -> UIManager | None:
+        """The UIManager this widget is attached to. During creation, this will be None."""
+        from arcade.gui.ui_manager import UIManager
+
+        w: UIWidget | None = self
+        while w and w.parent:
+            parent = w.parent
+            if isinstance(parent, UIManager):
+                return parent
+
+            w = parent
+        return None
 
     @property
     def left(self) -> float:
@@ -549,6 +590,8 @@ class UIInteractiveWidget(UIWidget):
         interaction_buttons: defines, which mouse buttons should trigger
             the interaction (default: left mouse button)
     """
+
+    focus_mode = FocusMode.ALL
 
     # States
     hovered = Property(False)
@@ -867,3 +910,6 @@ class UISpace(UIWidget):
     @color.setter
     def color(self, value):
         self.with_background(color=value)
+
+
+__all__ = ["Surface", "UIDummy", "FocusMode", "UIInteractiveWidget", "UIWidget"]